+++ src/Synapse/Synapse/Commands/DataPlaneCommands/SparkStatement/GetAzureSynapseSparkStatement.cs
@@ -7,7 +7,7 @@ using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Synapse
 {
-    [Cmdlet(VerbsCommon.Get, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.SparkStatement)]
+    [Cmdlet(VerbsCommon.Get, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.SparkStatement, DefaultParameterSetName = GetSparkStatementsByIdParameterSetName)]
     [OutputType(typeof(PSSynapseSparkStatement))]
     public class GetAzureSynapseSparkStatement : SynapseCmdletBase
     {