+++ src/Synapse/Synapse/Commands/ManagementCommands/SqlPool/TestAzureSynapseSqlPool.cs
@@ -7,7 +7,8 @@ using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Synapse
 {
-    [Cmdlet(VerbsDiagnostic.Test, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.SqlPool)]
+    [Cmdlet(VerbsDiagnostic.Test, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.SqlPool, DefaultParameterSetName = TestByNameParameterSet)]
+    [OutputType(typeof(bool))]
     public class TestAzureSynapseSqlPool : SynapseCmdletBase
     {
         private const string TestByNameParameterSet = "TestByNameParameterSet";